src/requirements.txt
antlr4-python3-runtime==4.11.1
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
fonttools==4.55.0
kiwisolver==1.4.7
loguru==0.7.2
matplotlib==3.9.2
networkx==3.4.2
numpy==2.1.3
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
scipy==1.14.1
six==1.16.0
tzdata==2024.2
win32-setctime==1.1.0
